requirements.txt
openai
pandas
numpy
faiss-cpu
sentence-transformers
python-dotenv